Fixes a bug in the client/server implementation
The recv() system call returns any data that is available, up to the
requested amount, rather than waiting for receipt of the full amount of
data requested. The previous implementation attempted to split messages
into sections using a ',', but this is a poor implementation due to the
assumptions that were made about receipt of data.

The fix adds two methods to the Client class, namely recv_all() and
recv_msg(). We use a struct in order to send the length of the data as a
message header (in network byte order). That struct is unpacked on the
client's end. recv_msg() first recvs and unpacks the 4 byte msg header
that tells us how many bytes we are supposed to recv. Finally,
recv_msg() calls recv_all() with that number of bytes and keeps recving
until the buffer is filled with `size` number of bytes. The
data_receiver() method was simplified by calling recv_mesg() and then
decoding, parsing, and queueing the messages.

The diagram in the README was rereated to paint a clearer picture of the
architecture.

Added the GPL3 header to the pyqt5 example.